+++ lib/adassault/dns.rb
@@ -1,3 +1,4 @@
 # frozen_string_literal: true
 
+require 'adassault/dns/duzdu'
 require 'adassault/dns/find_dcs'