+++ docs/CHANGELOG.md
@@ -2,7 +2,11 @@
 
 ## [Unreleased]
 
+## [0.0.2]
+
+- Add `DNS::DUZDU` class, DNS unsecure zone dynamic update
+
 ## [0.0.1]
 
 - First version (early stage)
-- Replaces [DCDetector](https://github.com/noraj/DCDetector) and integrates it under the `dns find_dcs` command.
+- Replaces [DCDetector](https://github.com/noraj/DCDetector) and integrates it under the `dns find_dcs` command (`DNS::FindDCs` class).